restructuring directory
